Error: Need to fix api null return values

+++ app/src/main/java/com/example/bodybuilder/data/DailyCalorieData/DailyCalorieData.kt
@@ -1,6 +1,6 @@
 package com.example.bodybuilder.data.DailyCalorieData
 
 data class DailyCalorieData(
-    val BMR: Int?,
+    val BMR: Float?,
     val goals: DailyCalorieGoalsData?
 )
